move pagination logic to the repo
