+++ BookLibrary-cdi/backend/src/main/java/com/softserveinc/booklibrary/backend/repository/EntityRepository.java
@@ -3,6 +3,7 @@ package com.softserveinc.booklibrary.backend.repository;
 import java.io.Serializable;
 import java.util.List;
 
+import com.softserveinc.booklibrary.backend.dto.paging.MyPage;
 import com.softserveinc.booklibrary.backend.entity.AbstractEntity;
 
 public interface EntityRepository<T extends AbstractEntity<? extends Serializable>> {
@@ -18,4 +19,6 @@ public interface EntityRepository<T extends AbstractEntity<? extends Serializabl
 	boolean isEntityValid(T entity);
 
 	List<T> getAll();
+
+	MyPage<T> listEntities(int pageId, int numEntitiesOnPage);
 }
