+++ BookLibrary-cdi/backend/src/main/java/com/softserveinc/booklibrary/backend/service/impl/AbstractEntityService.java
@@ -73,45 +73,7 @@ public abstract class AbstractEntityService<T extends AbstractEntity<? extends S
 	@Override
 	@Transactional(propagation = Propagation.SUPPORTS)
 	public MyPage<T> listEntities(int pageId, int numEntitiesOnPage) {
-		MyPage<T> page = new MyPage<>();
-		List<T> allEntities = getAll();
-		int totalElements = allEntities.size();
-		int totalPages = totalElements != 0 ? (int) Math.ceil((double) totalElements / numEntitiesOnPage) : 1;
-		page.setTotalElements(totalElements);
-		page.setTotalPages(totalPages);
-		if (pageId + 1 > totalPages || pageId < 0) {
-			pageId = 0;
-		}
-		setFirstLastNumElements(pageId, numEntitiesOnPage, page);
-		page.setPageable(new MyPageable(numEntitiesOnPage, pageId));
-		page.setContent(getAll().stream()
-				.skip(page.getNumberOfFirstPageElement() - 1)
-				.limit(numEntitiesOnPage).collect(Collectors.toList()));
-		return page;
-	}
-
-	private void setFirstLastNumElements(int pageId, int numEntitiesOnPage, MyPage<T> page) {
-		if (page.getTotalPages() == 1) {
-			page.setFirst(true);
-			page.setLast(true);
-			page.setNumberOfFirstPageElement(1);
-			page.setNumberOfElements(page.getTotalElements());
-		} else if (pageId == 0) {
-			page.setFirst(true);
-			page.setLast(false);
-			page.setNumberOfFirstPageElement(1);
-			page.setNumberOfElements(numEntitiesOnPage);
-		} else if (pageId + 1 == page.getTotalPages()) {
-			page.setLast(true);
-			page.setFirst(false);
-			page.setNumberOfFirstPageElement(pageId * numEntitiesOnPage + 1);
-			page.setNumberOfElements(page.getTotalElements());
-		} else {
-			page.setFirst(false);
-			page.setLast(false);
-			page.setNumberOfFirstPageElement(pageId * numEntitiesOnPage + 1);
-			page.setNumberOfElements((pageId + 1) * numEntitiesOnPage);
-		}
+		return repository.listEntities(pageId, numEntitiesOnPage);
 	}
 
 }
